Collections.namedtuple()

+++ 13.py
@@ -8,7 +8,7 @@ for i in range(N):
     students = namedtuple('student',fields)
     field1, field2 = input("Enter name and grade: ").split()
     student = students(field1,field2)
-    total += int(student.grade)
+    total += float(student.grade)
 print('{:.2f}'.format(total/N))
